fix(trainers): make the non-finite loss guard dtype-safe and rank-safe
The guard reduced the loss with `loss.detach().mean()`. A low-precision dtype has no `mean` kernel,
so a `float8_e5m2` loss raised `NotImplementedError: "sum_cpu" not implemented for 'Float8_e5m2'`
rather than reporting a rate. That turned a diagnostic into a crash on the dtypes most likely to
produce a non-finite loss in the first place. Casting to float32 before the reduction fixes both
float8 variants and leaves the finiteness of every other dtype unchanged. Applied at all 31 sites.

The guard also ends each loss path in a collective, and ten raises upstream of it ran on one rank
only. A rank that raised left its peers waiting inside a gather that never completed: measured on
two ranks, the peer blocked 3 times out of 3. Those ten now agree across ranks before raising, so
the run fails together instead of hanging. Four config-time raises stay rank-local on purpose,
because a gather inside a branch only the raising rank enters would create the hang it prevents
everywhere else.

BCO stored the metric under `is_main_process`, discarding every other rank's contribution to a
value the gather had already averaged across ranks. Its store line now matches CPO's and ORPO's.

Two tests were passing against a mutation. The SFT rate test split one non-finite micro-batch
against one finite one, so an inverted `~torch.isfinite` still averaged to 0.5 and the assertion
could not tell the two apart; it now uses four micro-batches and asserts 0.25. The DPO Liger loss
path had no test at all, so deleting its guard changed nothing observable. Both mutations now fail.

The 15 doc pages said the filter substitutes a finite value unconditionally. It does that only when
`is_torch_xla_available()` is false; under XLA the step is dropped instead.

+++ trl/experimental/minillm/minillm_trainer.py
@@ -418,7 +418,9 @@ class MiniLLMTrainer(GRPOTrainer):
         # the step that failed. Report the condition here instead. Gather first, so a rank whose loss went non-finite
         # is counted even when the other ranks are finite.
         mode = "train" if self.model.training else "eval"
-        nonfinite = self.accelerator.gather((~torch.isfinite(loss.detach().mean())).float())
+        # Cast before the reduction: a low-precision dtype such as `float8_e5m2` has no `mean` kernel, and the
+        # cast leaves the finiteness of every other dtype unchanged.
+        nonfinite = self.accelerator.gather((~torch.isfinite(loss.detach().float().mean())).float())
         self._metrics[mode]["frac_nonfinite_loss"].append(nonfinite.mean().item())
         if nonfinite.any():
             # `logging_nan_inf_filter` and the optimizer step belong to the training loop only, so each mode gets its
